ecommerce_price_tracker/pyproject.toml
[project]
name = "ecommerce-price-tracker"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "requests>=2.32.5",
    "streamlit>=1.52.2",
]
